chore(architecture): relocate module composition
- move composition beside owning modules and shared concerns

- expose shared technical APIs through focused indexes

- place HTTP routes under module interface layers
