+++ src/app/authentication/interface/http/login-handler.ts
@@ -1,8 +1,8 @@
 import { type HttpScope } from '@koala-ts/framework';
 import { type LoginRequest } from '@/app/authentication/interface/http/login-request';
 import { loginResponse } from '@/app/authentication/interface/http/login-response';
-import { login } from '@/composition/auth/login';
-import { mapResultToHttp } from '@/composition/shared/http/normalization';
+import { login } from '@/app/authentication/authentication-composition';
+import { mapResultToHttp } from '@/app/shared/interface/http';
 
 export async function loginHandler({ response, request }: HttpScope): Promise<void> {
   const result = await login(request.body as LoginRequest['body']);
