+++ src/app/home/interface/http/home-routes.ts
@@ -1,0 +1,4 @@
+import { Get } from '@koala-ts/framework/routing';
+import { homeHandler } from '@/app/home/interface/http/home-handler';
+
+export const homeRoute = Get('/', 'home', homeHandler);
