+++ src/app/shared/infrastructure/logger/index.ts
@@ -1,0 +1,1 @@
+export { appLogger } from './logger-composition';
