+++ src/app/shared/infrastructure/logger/logger-composition.ts
@@ -1,6 +1,6 @@
 import { type Logger } from '@/app/shared/application/logger';
 import { createLogger } from '@/app/shared/infrastructure/logger/pino';
-import { loggerConfig } from '@/config';
+import { loggerConfig } from '@/config/logger';
 
 export const appLogger: Logger = createLogger({
   level: loggerConfig.level,
