+++ src/app/shared/infrastructure/persistence/mongodb/index.ts
@@ -1,0 +1,3 @@
+export { mongoDBClient, usersCollection, workspacesCollection } from './mongodb-composition';
+export type { UsersCollection } from './users-collection-schema';
+export type { WorkspacesCollection } from './workspaces-collection-schema';
