+++ src/app/shared/interface/http/index.ts
@@ -1,2 +1,22 @@
-export * from './types';
-export * from './validation-middleware';
+export { mapResultToHttp, validateRequest } from './http-composition';
+export {
+  HTTP_BAD_REQUEST,
+  HTTP_CONFLICT,
+  HTTP_CREATED,
+  HTTP_INTERNAL_SERVER_ERROR,
+  HTTP_OK,
+  HTTP_UNAUTHORIZED,
+} from './status-code';
+export type { StatusCode } from './status-code';
+export type {
+  HttpMiddleware,
+  HttpRequest,
+  HttpResponse,
+  HttpScope,
+  NextMiddleware,
+  ValidationRules,
+  ValidationViolation,
+  Validator,
+  ViolationMapper,
+} from './types';
+export { validationMiddleware } from './validation-middleware';
