+++ src/app/workspace/interface/http/create-workspace-handler.ts
@@ -1,8 +1,8 @@
 import { type HttpScope } from '@koala-ts/framework';
 import { type CreateWorkspaceRequest } from '@/app/workspace/interface/http/create-workspace-request';
 import { createWorkspaceResponse } from '@/app/workspace/interface/http/create-workspace-response';
-import { mapResultToHttp } from '@/composition/shared/http/normalization';
-import { createWorkspace } from '@/composition/workspace/create-workspace';
+import { mapResultToHttp } from '@/app/shared/interface/http';
+import { createWorkspace } from '@/app/workspace/workspace-composition';
 
 export async function createWorkspaceHandler({ response, request }: HttpScope): Promise<void> {
   const result = await createWorkspace(request.body as CreateWorkspaceRequest['body']);
