+++ src/app/workspace/workspace-composition.ts
@@ -1,9 +1,9 @@
+import { workspacesCollection } from '@/app/shared/infrastructure/persistence/mongodb';
 import { createSaveWorkspace } from '@/app/shared/infrastructure/persistence/repository/workspace-repository/save-workspace';
+import { validateRequest } from '@/app/shared/interface/http';
 import { createCreateWorkspace } from '@/app/workspace/application/create-workspace';
 import { type SaveWorkspace } from '@/app/workspace/domain/workspace-repository';
 import { createWorkspaceRequestConstraints } from '@/app/workspace/interface/http/create-workspace-request';
-import { validateRequest } from '@/composition/shared/http/middleware';
-import { workspacesCollection } from '@/composition/shared/persistence/mongodb';
 
 const saveWorkspace: SaveWorkspace = createSaveWorkspace({ workspacesCollection });
