--- src/composition/shared/http/middleware.ts
@@ -1,5 +1,0 @@
-import { flattenViolations } from '@koala-ts/framework/validator';
-import { validationMiddleware } from '@/app/shared/interface/http';
-import { validate } from '@/composition/shared/http/validation';
-
-export const validateRequest = validationMiddleware(validate, flattenViolations);
