--- src/composition/shared/http/validation.ts
@@ -1,3 +1,0 @@
-import { builtInConstraints, createValidator } from '@koala-ts/framework/validator';
-
-export const validate = createValidator({ constraints: builtInConstraints });
