+++ tests/__vitest__/integration-test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach } from 'vitest';
-import { mongoDBClient } from '@/composition/shared/persistence/mongodb';
+import { mongoDBClient } from '@/app/shared/infrastructure/persistence/mongodb';
 
 export function integrationTest(): void {
   beforeEach(async () => {
