+++ tests/workspace.test.ts
@@ -1,7 +1,7 @@
 import { createTestAgent } from '@koala-ts/framework';
 import { integrationTest } from '@tests/__vitest__/integration-test';
 import { describe, expect, test } from 'vitest';
-import { workspacesCollection } from '@/composition/shared/persistence/mongodb';
+import { workspacesCollection } from '@/app/shared/infrastructure/persistence/mongodb';
 import { appConfig } from '@/config';
 
 describe('Workspace feature test', () => {
